one hand for L mouse
made it so that the L hand can operate the mouse on it's own without input from other limbs lol. I swapped the L hand layout for the numpad and mouse layers, now the L had mouse keystrokes are accessed via the L pinky modifier that I use most often to access the R hand numpad.